tf backend onnx run works in debug mode
